Changes from PR #1374
