+++ apis/Google.Cloud.Diagnostics.Common/Google.Cloud.Diagnostics.Common/ErrorReporting/ContextExceptionLogger.cs
@@ -28,15 +28,14 @@ namespace Google.Cloud.Diagnostics.Common
         /// on GAE or GCE the project ID will be detected from the platform.</param>
         /// <param name="serviceName"> An identifier of the service, such as the name of the executable or job. Cannot be null.</param>
         /// <param name="version">Represents the source code version that the developer provided. Cannot be null.</param>
-        /// <param name="options">The error reporting options. Can be null.</param>
-        /// <returns></returns>
+        /// <param name="options">The error reporting options. Can be null, if null default options will be used.</param>
+        /// <returns>An <see cref="IContextExceptionLogger"/> for the given options.</returns>
         public static IContextExceptionLogger Create(string projectId, string serviceName,
             string version, ErrorReportingOptions options)
         {
             GaxPreconditions.CheckNotNull(serviceName, nameof(serviceName));
             GaxPreconditions.CheckNotNull(version, nameof(version));
-            return ErrorReportingContextExceptionLogger.Create(
-                projectId, serviceName, version, options);
+            return ErrorReportingContextExceptionLogger.Create(projectId, serviceName, version, options);
         }
     }
 }
