+++ apis/Google.Cloud.Diagnostics.Common/Google.Cloud.Diagnostics.Common/ErrorReporting/EventTarget.cs
@@ -74,7 +74,7 @@ namespace Google.Cloud.Diagnostics.Common
         public static EventTarget ForLogging(string projectId = null, string logName = LogNameDefault,
             LoggingServiceV2Client loggingClient = null, MonitoredResource monitoredResource = null)
         {
-            projectId = CommonUtils.GetAndCheckProjectId(projectId, monitoredResource);
+            projectId = Project.GetAndCheckProjectId(projectId, monitoredResource);
             var logTarget = LogTarget.ForProject(projectId);
             return ForLogging(logTarget, logName, loggingClient, monitoredResource);
         }
